Part18
• Functional copy and delete buttons on Gist View

+++ lib/elixir_gist/gists.ex
@@ -4,6 +4,7 @@ defmodule ElixirGist.Gists do
   """
 
   import Ecto.Query, warn: false
+  alias ElixirGist.Accounts.User
   alias ElixirGist.Repo
 
   alias ElixirGist.Gists.Gist
@@ -86,8 +87,15 @@ defmodule ElixirGist.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_gist(%Gist{} = gist) do
-    Repo.delete(gist)
+  def delete_gist(%User{} = user, gist_id) do
+    gist = Repo.get!(Gist, gist_id)
+
+    if user.id == gist.user_id do
+      Repo.delete(gist)
+      {:ok, gist}
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
